Fixed page4. Altered page6.
