+++ Spa.test/main.js
@@ -1,9 +1,9 @@
 import { render as page1Render } from "./home.js";
 import { render as page2Render } from "./page2.js";
 import { render as page3Render } from "./page3.js";
-import { render as page4Render } from "./page4.js";
-import { render as page5Render } from './filters.js';
 import { render as ListAllExpensesInBudgetRender } from "./ListAllExpensesInBudget.js";
+import { render as page5Render } from './filters.js';
+import { render as ListAllExpensesInBudgetCatRender } from "./ListAllExpensesInBudgetCat.js";
 import { registerChangeListener } from "./router.js";
 import { Header } from "./header.js";
 
@@ -22,14 +22,14 @@ registerChangeListener((newPage) => {
         case 'page3':
             page3Render(app);
             break;
-        case 'page4':
-            page4Render(app);
+        case 'ListAllExpensesInBudget':
+            ListAllExpensesInBudgetRender(app);
             break;
         case 'filters':
             page5Render(app)
             break;
-        case 'ListAllExpensesInBudget':
-            ListAllExpensesInBudgetRender(app);
+        case 'ListAllExpensesInBudgetCat':
+            ListAllExpensesInBudgetCatRender(app);
             break;
     }
 })
